Defer bin/link to Bullet Train's bt:link task
Since Bullet Train already ships with a task, it should know how to link all our appropriate gems adequately,
and we don't need to burden apps as much with this.

We don't expect the task to be executed with `bundle exec` so we keep bin/link around.

See https://github.com/bullet-train-co/bullet_train-core/pull/141