Directly check if node is installed in bin/configure
Previously, users who install node with a package manager other than
homebrew would receive a message that the package wasn't installed
via homebrew. Now the code checks whether node is present AND whether
it is the correct version in the same place.

There is a subtle difference in that the previous version would allow
you to continue without installing node but I think this was likely not
intended.