Bug fix for #615
I noticed after I merged #615 (😬 🤦‍♂️) that the tests in `main`
were failing. It turns out that having the assignments for
`@menu_orientation` inside of the `capture` block made it so that we
didn't capture the `render` call, but instead we seem to capture a
`nil`. I think this is because that final assignment returns nothing to
the template. This moves the assignemtns outside of the `capture` block.