Run tests against release gems and `main` in `core`
This makes it so that every PR and every merge to main will run tests in
CI against both the released versions of the `core` gems and against a
branch in the `core` repo.

For merges to `main` in the starter repo, they'll always run against the
latest state of `main` in `core`.

For PRs we'll attempt to find a branch in `core` that matches the name
of the branch being tested. If we can find one, we'll test against that.
If not we'll test the PR branch against `main` in `core`.

The tests that we run against the released versions of the gems can
probably be expected to fail with some regularity since we sometimes
introduce changes into the starter repo that depend on particular
changes in `core`. But this will help us to know when we've made changes
in the starter repo that mean that we need to relase a new version of
the `core` gems.