Add tests for Stripe webhooks (#881)
This implements some tests for bullet-train-pro/bullet_train-billing-stripe#18. If the bullet_train-billing-stripe gem is not installed then these new tests will be skipped.

It also removes a (seemingly unused) factory that was duplicated out of bullet_train-billing-stripe.

This app still has access to that factory via the changes made to test_helper.rb.