Add GitHub workflow for upgrading Bullet Train (#1077)
* Add GitHub workflow for updating the Bullet Train version

Fixes https://github.com/bullet-train-co/bullet_train/issues/978

This is my first stab at making a workflow that will generate a PR for
updating an app against the Bullet Train starter repo.

In order to generate the PR we're using the `theirs` resolution method
to deal with any conflicts. In this context that means that if a block
of code has been changed in the downstream app _and_ in the starter
repo, then the PR that we generate will contain a change set that will
clobber the changes in the downstream repo with the change that was made
in the starter repo.

For instance, my app was on 1.5.2, and I changed this line:

https://github.com/bullet-train-co/bullet_train/blob/2ea5c9c136568bae39257af6c8944267b058dc17/test/system/invitation_lists_test.rb#L29

to read:

```ruby
assert page.has_content?("Tell")
```

And then when I ran this action in my repo the resulting PR proposed
this change:

```diff
- assert page.has_content?("Tell")
+ assert_text("Tell us about you")
```

That change was just buried in amongst all the other changes in the PR
and would be very easy to miss during a review.

Currently this workflow requires that a Personal Access Token be created
and set in the repo secrets as `PAT`. The token can be a fine-grained
token as long as it has "Contets" and "Pull Request" permissions set to
read/write.

I opted to use a PAT because of limitations on the default token that is
used for GitHub actions. Any workflows using the default token can't
trigger new workflows as side effects. Specifically for this action it
means that CI workflows for the resulting PR are not run when using the
default token.

* Update .github/workflows/update-bullet-train.yml

Co-authored-by: Kasper Timm Hansen <kaspth@gmail.com>

* Pull in updates from seshy-rails

* Use diff3

* Use an external action

* Use the work 'upgrade' instead of 'update'

---------

Co-authored-by: Kasper Timm Hansen <kaspth@gmail.com>