Fix for new versions of Chrome/webdriver (#1243)
* Trying to debug a flaky test

* will this help

* Fix for new versions of Chrome/webdriver

See this issue on Capybara for additional details:
https://github.com/teamcapybara/capybara/pull/2726

This PR just registers a new version of `:selenium_chrome_headless`
instead of using the one registered in the `capybara` gem. The new one
just mirrors the fix that's already been applied to the gem, but not yet
released to rubygems.org. Once a new version of `capybara` is out, we
can remove the custom driver.

* dang linter

* Don't actually need that