Redirect `device_test` definitions to correct test file.
All this is because Rails' `test` method uses `define_method`, which will report this file as the `source_location`
and not the target test file where the test is actually defined.

`define_method` cannot be passed the correct location, we must use `class_eval` with a source string of Ruby for that.
But using a string means we can't refer to the block, like `define_method(name, &block)` would.

So we inject a module that we define methods via `test` to pass the block to, then we override that
via `class_eval` to point it to the correct location.