Fix Heroku deploys for yarn 4
Heroku recommends having the `heroku/nodejs` buildpack before
`heroku/ruby`. So I updated `app.json` to to that.

But the `heroku/nodejs` buildpack tries to run `yarn build`
automatically. But our build steps shell out to bundler, and since
`heroku/ruby` comes after `heroku/nodjs` bundler isn't yet available.

Adding a `heroku-postbuild` script makes it so that Heroku runs _that_
script instead of running the normal `build` script. Which prevents the
lack of ruby/bundler from being a problem during the setup phase for the
`heroku/nodejs` buildpack.

Then to ensure that `yarn build` is run, I've added that to the end of
the `release` phase in `Procfile`.