Make it easier to use Cloudinary as a storage option for ActiveStorage
This makes it easier to use Cloudindary as the backend service for
ActiveStorage. From a functional perspective this works very much like
the S3 service option. ActiveStorage is still in control of files, but
it sends them through to Cloudinary for storage and delivery.

Partial fix for : https://github.com/bullet-train-co/bullet_train/issues/1617