Make it easier to use Cloudinary as a storage option for ActiveStorage (#1739)
* Make it easier to use Cloudinary as a storage option for ActiveStorage

This makes it easier to use Cloudindary as the backend service for
ActiveStorage. From a functional perspective this works very much like
the S3 service option. ActiveStorage is still in control of files, but
it sends them through to Cloudinary for storage and delivery.

**NOTE:** If, in production, you use Cloudinary for anything and have a `CLOUDINARY_URL` ENV var set but you want to use the `:local` option for ActiveStorage then you'll want to make an adjustment to `config/environments/production.rb`.

Partial fix for : https://github.com/bullet-train-co/bullet_train/issues/1617