Refactor check_image_dependencies.rb to not require both vips and ImageMagick (#1765)
* Refactor check_image_dependencies.rb to not require both vips and ImageMagick

This makes it so that if you have either `vips` or `magick` available on
your `PATH` that we count that as being good enough for image processing
dependencies and we move on. Previously we'd check for both and prompt
you to install both even though we only need one or the other. This
should make it less confusing for people who don't realize that they're
alternatives that do the same thing.

* linter