Improve Heroku Detection
We can use app.json to set an environment variable for apps deployed in
Heroku. This is kind of the "first line of defense" for detecting that
we're in Heroku. We'll also fallback to some other methods.

When the `BT_IS_IN_HEROKU` ENV var is set to `true` we'll assume that
we're running in Heroku, and we'll sidestep any other detection methods.

We'll still fallback to our previous detection methods if that ENV var
does not exist.

This also allows apps _not_ in Heroku to negate a false positive
detection by our fallback methods by setting the `BT_IS_IN_HEROKU` ENV
var to `false`.