Add `microscope` as an optional gem
We used to list `microscope` in the `.gemspec` for the `bullet_train`
gem, which forced it to be present in _all_ Bullet Train apps. To try to
reduce memory use we're now including it as a "recommended bu optional"
gem that people can opt-in to using.

Joint PR: https://github.com/bullet-train-co/bullet_train-core/pull/1035