Add `ruby-openai` as an optional gem (#1927)
We used to include `ruby-openai` as a dependency in the `.gemspec` of
the main `bullet_train` gem. But we don't actually use it in any of the
`core` gems or in the starter repo directly. The only place that we use
it is in the `bullet_train-action_models` gem, and even there it's an
optional dependency.

If you want to continue to use it you can uncomment the `ruby-openai`
line in the `Gemfile`.

Joint PR: https://github.com/bullet-train-co/bullet_train-core/pull/1037