Add `awesome_print` as an optional gem
We used to list `awesome_print` as a hard dependency in the main
`bullet_train` gem. In an effort to reduce memory use in a new BT app
we're now adding it as an optional dependency that developers can opt-in
to using if they want to.

Joint PR: https://github.com/bullet-train-co/bullet_train-core/pull/1042