Move `pry` into the `:development, :test` gem group (#1929)
Previously we had `pry` in all groups, which meant that we were shipping
it to production, but really you shouldn't need a debugger in production
unless you really know what you're doing. Moving it to the
`:development, :test` group makes it available in the places that you're
most likely to need it while slimming down the production footprint.

If you need to use it in production for your app you can move it out of
the `:development, :test` group in the `Gemfile`.