Make a bunch of production gems optional
We used to include several gems in the `:production` group of the
`Gemfile` that may not be used in downstream apps. This PR comments them
out in the `Gemfile` with the idea that people who need to use them can
uncomment them to opt-in. This will make the default deployment image
smaller than it was previously.