Hoist `sidekiq` into the `Gemfile` of the starter repo
Previously we had listed `sidekiq` as a hard dependency of the `core`
gems. This (and the joint PR) make it so that we list `sidekiq` directly
in the `Gemfile` of the starter repo instead of as a gem dependency of
the `core` gems. That will allow people to remove sidekiq if they want
to use something else for background processing.