Override the `action_text:install` task
We already have most of the `ActionText` stuff in the `core` repo, and
are considering moving the remaining bits there as well. We're
overriding this task to prevent confusion if someone tries to run the
installation process. We're including a `--clobber` option so that
people can still run the original task if they really want to.